+++ lal/linarr/D.cpp
@@ -60,9 +60,9 @@ namespace linarr {
 
 /* D */
 
-inline uint64_t __sum_length_edges(const graph& g, const LINARR& pi) {
+inline uint32_t __sum_length_edges(const graph& g, const LINARR& pi) {
 	// sum of lengths
-	uint64_t l = 0;
+	uint32_t l = 0;
 
 	edge_iterator it(g);
 	while (it.has_next()) {
@@ -77,7 +77,7 @@ inline uint64_t __sum_length_edges(const graph& g, const LINARR& pi) {
 	return l;
 }
 
-uint64_t sum_length_edges(const graph& g, const LINARR& pi) {
+uint32_t sum_length_edges(const graph& g, const LINARR& pi) {
 	assert(pi.size() == 0 or g.n_nodes() == pi.size());
 	return utils::call_with_empty_arrangement(__sum_length_edges, g, pi);
 }
@@ -85,7 +85,7 @@ uint64_t sum_length_edges(const graph& g, const LINARR& pi) {
 /* MDD */
 
 inline rational __MDD_rational(const graph& g, const LINARR& pi) {
-	const uint64_t D = sum_length_edges(g, pi);
+	const uint32_t D = sum_length_edges(g, pi);
 	return rational_from_ui(D, g.n_edges());
 }
